--- a/babelTransform.js
+++ b/babelTransform.js
@@ -5,15 +5,14 @@ module.exports = {
     const alias = options?.config?.moduleNameMapper;
     const extensions = options.config.moduleFileExtensions;
     const babelTransformer = babelJest.createTransformer({
-     presets: [
-        ["@babel/preset-react", { runtime: "automatic" }], // Ensuring JSX transform
-        ["@babel/preset-env"],
-        ["@babel/preset-typescript", { isTSX: true, allExtensions: true }] // Adjusted for TSX
-      ],
-      plugins: [["transform-barrels", { executorName: "jest", alias: alias, extensions: extensions }]],
+      presets: [["@babel/preset-react"], ["@babel/preset-env"], ["@babel/preset-typescript"]],
+      plugins: [
+        ["transform-barrels", { executorName: "jest", alias: alias, extensions: extensions, isCacheEnabled: false }],
+      ], // set to true when all else worked
       babelrc: false,
       configFile: false,
     });
+
     return babelTransformer.process(sourceText, sourcePath, options);
   },
 };