--- a/package.json
+++ b/package.json
@@ -25,6 +25,7 @@
     "acorn-jsx-walk": "^2.0.0",
     "acorn-walk": "^8.3.2",
     "arg": "^5.0.2",
+    "babel-plugin-transform-barrels": "github:cmfrydos/babel-plugin-transform-barrels",
     "bcryptjs": "^2.4.3",
     "better-react-mathjax": "^2.0.3",
     "clsx": "^1.2.1",