… test/jest/config/babelTransform.js