On request from FicocelliGuy in development: https://discord.com/channels/415207508303544321/459097632896188436/1280207617062469662 

This is just fiddling arround with [babel-plugin-transform-barrels](https://github.com/cmfrydos/babel-plugin-transform-barrels)
following this guide: https://dev.to/fogel/potential-issues-with-barrel-files-in-jest-1nkl
it currently has many problems:
- I struggle to put babelTransform.js anywhere else than the root directory. (Edit: fixed)
- Tests now only work from the root-directory (working directory cannot be /src or /test) (Edit: only for jest command)
- Using the debugger in Visual Studio Code is broken now. (Could later not be verified)

So this needs some work to fix these issues first.
I suspect the first two being related to false pathing, and the second one due to changing line numbering.

It's faster though:
Here are some test-performance-tests running a 5950X with 32GB (18GB free).
I excluded test\jest\Netscript\Bladeburner.test.ts since it's running stupid amounts of testcases (10 000 000), which just take a lot of time on their own. So to meassure jest startup time, which is the real problem here, I excluded it.

Before (all tests) : 50.644s / 15GB usage
![grafik](https://github.com/user-attachments/assets/9f7f2cce-7aab-48ae-9036-e26066e53954)

After (all tests): 15.721s / 10GB usage
![grafik](https://github.com/user-attachments/assets/14d0ffe7-159b-4cae-8b5f-3f61cc5e9024)

Ram-Screenshot
![grafik](https://github.com/user-attachments/assets/eb005dd3-4047-4c4e-bb47-1df417a24403)

After (all tests with --maxWorkers=4 param): 16.18 s / 4GB usage
![grafik](https://github.com/user-attachments/assets/ec24ca34-8c1a-45a0-9204-fc8bd4854056)
**This might be the ideal config for people with 16GB or less RAM.**

Before (one test): 20.835s (Corporation.test.ts)
![grafik](https://github.com/user-attachments/assets/5a2e8f35-2308-4e16-aabe-81b253e4ec65)

After (one test): 6.803s (Corporation.test.ts)
![grafik](https://github.com/user-attachments/assets/a6d09db8-944a-43f4-8643-6b892b29438d)

Further ToDo:
- [x] check whether opening a subbash in root directory when calling `npm run test` will fix the plugins pathing issues.
(That's actually package.jsons default behaviour)
- [x] investigate the ast node generation and why the source maps are not generated properly
- [x] ~~try make the plugin work with Webpack to potentially reduce output bundle size and improve performance~~ No need.