@catloversg 2024-09-03
You must not exclude any test. If you want to tune a test, you have to make a separate PR.