@cmfrydos 2024-09-08
ah, the tests failing is not a good sign. It worked at my computer, even without caching. I'll investigate this though tomorrow.

Also, I couldn't really fix the problem that the jest command now fails at any directory but the root dir. I've tried, and removed many process.cwd calls, or ones to path.transform("."). Calling npm run test works everywhere though.

The plugin could potentially also reduce the webpack size and improve general game performance. I wanted to test this, but somehow this also introduces new errors I didn't worked through.

The last point to check would be to assure the source-maps are created correctly such that debugging testruns would be possible (again).

@d0sboots 
And would you trust this plugin enough, including my changes to it, to let it transform all our tests and codebase? It may also introduce failing tests in the future, just because some obscure js syntax is throwing it off. For me, these drawbacks aren't really worth it for the 2-3x speedup, atleast for others who did not experiment with it as I did the last days and knows potential error sources.