@d0sboots 2024-09-08
I'm not going to trust some random thing that isn't even published as an npm module (which anyone can do) but instead is just someone's random github blob, to operate as part of our release process. There's also significantly less need; the speed of building is not a bottleneck there, and all that matters is the final produced output. By contrast, running tests is much more about speed, and the size of the output is not a concern since it will be thrown away after the tests run.

It *is* important that all the tests be able to run, but if there's a minor issue we can work around it. That's not the issue that I saw in the test logs though, it seemed like something much more systematic (but hopefully easy to fix).