@cmfrydos 2024-09-08
To be precise, it is a npm module, but I needed to fork it to make it compatible with this codebase. 
(Maybe you should fork it from me after all issues are resolved, so that I could not install any backdoor on your testing setup 😉. 

It is a rather "unknown" babel plugin though, looking at the amount of issues of the original repo, which explains the troubles it has with a larger codebase covering some more unlikely edge cases.