@cmfrydos 2024-09-08
The tests should pass now. I accidentally pushed [babel-plugin-transform-barrels](https://github.com/cmfrydos/babel-plugin-transform-barrels) in a broken state earlier. This likely happened because I first edited the plugin files directly and only updated the fork afterward.

Now, I just need to fix the AST source-map generation for this plugin.

One last point to clarify: should I modify the transformer plugin to only 'activate' on the `npm run test:fast` command and leave the npm test command as it was? Or would it be acceptable to make this faster version the new standard test procedure, with the drawback that tests may unexpectedly fail when, for example, a new module is installed that uses additional unused JS import/export syntax? In that case, you'd encounter an error similar to the one in the last failed test: https://github.com/bitburner-official/bitburner-src/pull/1645/commits/63d696fadb185ddd6947554284c59e0f5013527f