@d0sboots 2024-09-08
> One last point to clarify: should I modify the transformer plugin to only 'activate' on the `npm run test:fast` command and leave the npm test command as it was? Or would it be acceptable to make this faster version the new standard test procedure, with the drawback that tests may unexpectedly fail when, for example, a new module is installed that uses additional unused JS import/export syntax?

This should work with `npm test`, and we should not introduce a new `npm test:fast`. There needs to be only one way to run tests, and it needs to be able to run all the tests. But it's fine if it introduces (reasonable) restrictions on the ways we can write new tests.