@d0sboots 2024-09-08
As a separate note, the success of this makes me think we should look into refactoring for removal of barrel file usage. I don't think we have much of it in the BB codebase directly, but MUI has a lot of it and it's completely optional - we could import only what we use relatively easily.