@cmfrydos 2024-09-08
That's a good point. We should analyse the plugins output and compare it with the current implementation. I also prefer a more verbose, but detailed, import section more than just importing a wildcard. Maybe the plugin already reformatted the files in an appropriate way and there isn't much more manual to do. I'll come back here, when I tested that and fixed the remaining source-mapping issue (shouldn't be too hard).