@cmfrydos 2024-09-08
I've counted them! Those are all the files / import statements that use barrel import in bb:
Total .js/.ts/.tsx/.jsx files: 298
Total imports: 623

So this would be another +- 600 line refactor xD I can make a pr with all of them if you wish, but we can also move this to a later stage.