@cmfrydos 2024-09-08
Ok, then let's skip that for now.

I believe I've fixed the AST source map generation, but I'd like to verify this. Previously, I noticed some strange debugging behavior, such as breakpoints not stopping at the correct line. I'd like to replicate this with the last version to see if my recent changes have indeed fixed the issue. If so, we could go live with the minor drawback, as I previously mentioned, that the `jest` command will only work in the root directory and not in any subdirectories. However, `npm run test` will continue to work from anywhere, as before.