@cmfrydos 2024-09-08
Also, I'd like to cleanup my last push to the plugin-fork, and then we can set the version to this specific commit, such that it cannot be changed. I've already noticed the maintainer of the plugin about our issues with it, and hopefully my fixes will somehow go into their official code in a more robust way.