@cmfrydos 2024-09-09
Somehow, I couldn't replicate the debugging issues I had initially. I had struggled a lot with the caching at that time, and the disguised issues might have interfered with the debugging process. I read that including line information in the AST generation is recommended, which wasn't the case with this plugin. I've added this now, but I'm not sure if it was actually necessary.

I've pinned the version to the latest commit I made to the plugin. This is ready for merging from my end.

For anyone testing this, I recommend starting with the `jest --clear-cache` command to ensure that the new, faster transformer is used. Then, based on your system's memory:
- If you have 16GB of RAM or less, run `jest --maxWorkers=4` from the root directory.
- If memory is not a constraint, simply use `npm run test`.

The initial test run may be slow as Jest rebuilds its cache, but subsequent runs should be way faster.